# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: ISMS Builder API
  version: 1.0.0
  description: |
    REST API for the ISMS Builder platform — a self-hosted Information Security
    Management System (ISMS) according to ISO 27001:2022, NIS2, GDPR, BSI
    IT-Grundschutz and other frameworks.

    **Authentication:** All protected endpoints require a valid JWT cookie
    (`sm_session`). Obtain it via `POST /login`.

    **RBAC Roles (ascending rank):**
    - `reader` / `revision` (rank 1) — read access
    - `editor` / `dept_head` / `qmb` (rank 2) — create/edit
    - `contentowner` / `auditor` (rank 3) — approve, CISO/DSB settings
    - `admin` (rank 4) — user management, delete, system config

  contact:
    name: Claude Hecker
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
    description: Self-hosted instance
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sm_session
...
paths:
...
  /legal/summary:
    get:
      tags: [legal]
      summary: Legal summary (active contracts, expiring contracts)
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Legal summary
  /legal/contracts:
    get:
      tags: [legal]
      summary: List contracts
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Array of contracts
    post:
      tags: [legal]
      summary: Create contract (contentowner+)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Created contract
  /legal/contracts/expiring:
    get:
      tags: [legal]
      summary: List expiring contracts (within cancellation notice period)
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Array of expiring contracts
  /legal/contracts/{id}:
    get:
      tags: [legal]
      summary: Get single contract
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Contract
    put:
      tags: [legal]
      summary: Update contract (contentowner+)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
    delete:
      tags: [legal]
      summary: Soft-delete contract (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
...
  /legal/ndas:
    get:
      tags: [legal]
      summary: List NDAs
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Array of NDAs
    post:
      tags: [legal]
      summary: Create NDA (contentowner+)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Created NDA
  /legal/ndas/{id}:
    get:
      tags: [legal]
      summary: Get single NDA
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: NDA
    put:
      tags: [legal]
      summary: Update NDA (contentowner+)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
    delete:
      tags: [legal]
      summary: Soft-delete NDA (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
  /legal/policies:
    get:
      tags: [legal]
      summary: List privacy policies
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Array of privacy policies
    post:
      tags: [legal]
      summary: Create privacy policy (contentowner+)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Created policy
  /legal/policies/{id}:
    get:
      tags: [legal]
      summary: Get single privacy policy
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Policy
    put:
      tags: [legal]
      summary: Update privacy policy (contentowner+)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
    delete:
      tags: [legal]
      summary: Soft-delete privacy policy (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted

  # ─── TRAINING ─────────────────────────────────────────────────────────────
...